extract webpack entries for reuse
